backend/requirements.txt
 fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
scipy==1.11.4
scikit-image==0.22.0
PyWavelets==1.4.1
pymongo==4.5.0
cryptography==41.0.7
